Fixed guid bytes order in SMSG_AREA_SPIRIT_HEALER_TIME, CMSG_PETITION_DECLINE, CMSG_TURN_IN_PETITION. Added handling for GAMEOBJECT_TYPE_TEXT. Fixed SMSG_NPC_TEXT_UPDATE working with GameObject gossip. Found SMSG_GOSSIP_COMPLETE. More finishing touches to Deepwind Gorge.

+++ src/server/game/Handlers/MiscHandler.cpp
@@ -2262,7 +2262,7 @@ void WorldSession::HandleAreaSpiritHealerQueryOpcode(WorldPacket& recvData)
     recvData.ReadByteSeq(guid[6]);
     recvData.ReadByteSeq(guid[1]);
 
-    Creature* unit = GetPlayer()->GetMap()->GetCreature(guid);
+    Creature* unit = GetPlayer()->GetMap()->GetCreature(uint64(guid));
     if (!unit)
         return;
 
@@ -2270,10 +2270,10 @@ void WorldSession::HandleAreaSpiritHealerQueryOpcode(WorldPacket& recvData)
         return;
 
     if (bg)
-        sBattlegroundMgr->SendAreaSpiritHealerQueryOpcode(_player, bg, guid);
+        sBattlegroundMgr->SendAreaSpiritHealerQueryOpcode(_player, bg, uint64(guid));
 
     if (Battlefield* bf = sBattlefieldMgr->GetBattlefieldToZoneId(_player->GetZoneId()))
-        bf->SendAreaSpiritHealerQueryOpcode(_player, guid);
+        bf->SendAreaSpiritHealerQueryOpcode(_player, uint64(guid));
 }
 
 void WorldSession::HandleAreaSpiritHealerQueueOpcode(WorldPacket& recvData)
@@ -2302,7 +2302,7 @@ void WorldSession::HandleAreaSpiritHealerQueueOpcode(WorldPacket& recvData)
     recvData.ReadByteSeq(guid[0]);
     recvData.ReadByteSeq(guid[1]);
 
-    Creature* unit = GetPlayer()->GetMap()->GetCreature(guid);
+    Creature* unit = GetPlayer()->GetMap()->GetCreature(uint64(guid));
     if (!unit)
         return;
 
@@ -2310,10 +2310,10 @@ void WorldSession::HandleAreaSpiritHealerQueueOpcode(WorldPacket& recvData)
         return;
 
     if (bg)
-        bg->AddPlayerToResurrectQueue(guid, _player->GetGUID());
+        bg->AddPlayerToResurrectQueue(uint64(guid), _player->GetGUID());
 
     if (Battlefield* bf = sBattlefieldMgr->GetBattlefieldToZoneId(_player->GetZoneId()))
-        bf->AddPlayerToResurrectQueue(guid, _player->GetGUID());
+        bf->AddPlayerToResurrectQueue(uint64(guid), _player->GetGUID());
 }
 
 void WorldSession::HandleHearthAndResurrect(WorldPacket& /*recvData*/)
